CodeForgePro/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "eventlet>=0.39.0",
    "flask-socketio>=5.5.1",
    "flask==2.3.3",
    "joblib>=1.4.2",
    "matplotlib>=3.10.0",
    "numpy>=2.2.3",
    "oauthlib>=3.2.2",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.4",
    "python-binance>=1.0.27",
    "python-dotenv>=1.0.1",
    "python-socketio[client]>=5.12.1",
    "python-telegram-bot>=21.10",
    "scikit-learn>=1.6.1",
    "trafilatura>=2.0.0",
    "twilio>=9.4.6",
    "werkzeug==2.3.7",
    "simple-websocket>=1.0.0",
]
